+++ lib/src/widget/asset_picker.dart
@@ -7,10 +7,10 @@ import 'dart:math' as math;
 import 'dart:typed_data';
 import 'dart:ui' as ui;
 
-import 'package:flutter/material.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/services.dart';
 import 'package:extended_image/extended_image.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:photo_manager/photo_manager.dart';
 import 'package:provider/provider.dart';
 import 'package:wechat_assets_picker/wechat_assets_picker.dart';
@@ -38,7 +38,7 @@ class AssetPicker extends StatelessWidget {
         ),
         gridCount = gridCount ?? 4,
         themeColor =
-            pickerTheme?.colorScheme?.primary ?? themeColor ?? C.themeColor,
+            pickerTheme?.colorScheme?.secondary ?? themeColor ?? C.themeColor,
         super(key: key) {
     Constants.textDelegate = textDelegate ?? DefaultTextDelegate();
   }
@@ -212,6 +212,21 @@ class AssetPicker extends StatelessWidget {
           brightness: Brightness.dark,
           elevation: 0,
         ),
+        colorScheme: ColorScheme(
+          primary: Colors.grey[900],
+          primaryVariant: Colors.grey[900],
+          secondary: themeColor,
+          secondaryVariant: themeColor,
+          background: Colors.grey[900],
+          surface: Colors.grey[900],
+          brightness: Brightness.dark,
+          error: const Color(0xffcf6679),
+          onPrimary: Colors.black,
+          onSecondary: Colors.black,
+          onSurface: Colors.white,
+          onBackground: Colors.white,
+          onError: Colors.black,
+        ),
       );
 
   /// [ThemeData] for picker.
@@ -259,10 +274,10 @@ class AssetPicker extends StatelessWidget {
                         child: Transform.rotate(
                           angle: provider.isSwitchingPath ? math.pi : 0.0,
                           alignment: Alignment.center,
-                          child: const Icon(
+                          child: Icon(
                             Icons.keyboard_arrow_down,
-                            color: Colors.black,
                             size: 20.0,
+                            color: theme.colorScheme.primary,
                           ),
                         ),
                       ),
@@ -302,8 +317,8 @@ class AssetPicker extends StatelessWidget {
                       if (_.watch<AssetPickerProvider>().requestType ==
                           RequestType.audio) {
                         return ColoredBox(
-                          color: Colors.white12,
-                          child: Center(child: Icon(Icons.audiotrack)),
+                          color: theme.colorScheme.primary.withOpacity(0.12),
+                          child: const Center(child: Icon(Icons.audiotrack)),
                         );
                       }
 
@@ -319,7 +334,9 @@ class AssetPicker extends StatelessWidget {
                           fit: BoxFit.cover,
                         );
                       } else {
-                        return ColoredBox(color: Colors.white12);
+                        return ColoredBox(
+                          color: theme.colorScheme.primary.withOpacity(0.12),
+                        );
                       }
                     },
                   ),
@@ -343,8 +360,10 @@ class AssetPicker extends StatelessWidget {
                       ),
                       Text(
                         '(${pathEntity.assetCount})',
-                        style:
-                            TextStyle(color: Colors.grey[600], fontSize: 18.0),
+                        style: TextStyle(
+                          color: theme.textTheme.caption.color,
+                          fontSize: 18.0,
+                        ),
                         maxLines: 1,
                         overflow: TextOverflow.ellipsis,
                       ),
@@ -400,7 +419,7 @@ class AssetPicker extends StatelessWidget {
             child: Container(
               width: Screens.width,
               height: maxHeight,
-              decoration: BoxDecoration(color: theme.primaryColor),
+              decoration: BoxDecoration(color: theme.colorScheme.background),
               child: Selector<AssetPickerProvider,
                   Map<AssetPathEntity, Uint8List>>(
                 selector: (BuildContext _, AssetPickerProvider provider) =>
@@ -496,7 +515,7 @@ class AssetPicker extends StatelessWidget {
             ),
             onTap: () {
               if (provider.isSelectedNotEmpty) {
-                provider.successCallback(context, provider.selectedAssets);
+                Navigator.of(context).pop(provider.selectedAssets);
               }
             },
           );
@@ -515,7 +534,7 @@ class AssetPicker extends StatelessWidget {
             gradient: LinearGradient(
               begin: AlignmentDirectional.bottomCenter,
               end: AlignmentDirectional.topCenter,
-              colors: <Color>[Colors.black45, Colors.transparent],
+              colors: <Color>[theme.dividerColor, Colors.transparent],
             ),
           ),
           child: Align(
@@ -558,7 +577,7 @@ class AssetPicker extends StatelessWidget {
           gradient: LinearGradient(
             begin: AlignmentDirectional.bottomCenter,
             end: AlignmentDirectional.topCenter,
-            colors: <Color>[Colors.black45, Colors.transparent],
+            colors: <Color>[theme.dividerColor, Colors.transparent],
           ),
         ),
         child: Padding(
@@ -575,6 +594,12 @@ class AssetPicker extends StatelessWidget {
 
   /// Video asset type indicator.
   /// 视频类型资源指示
+  ///
+  /// Videos often contains various of color in the cover,
+  /// so in order to keep the content visible in most cases,
+  /// the color of the indicator has been set to [Colors.white].
+  /// 视频封面通常包含各种颜色，为了保证内容在一般情况下可见，此处
+  /// 将指示器的图标和文字设置为 [Colors.white]。
   Widget videoIndicator(AssetEntity asset) {
     return Align(
       alignment: AlignmentDirectional.bottomStart,
@@ -586,18 +611,26 @@ class AssetPicker extends StatelessWidget {
           gradient: LinearGradient(
             begin: AlignmentDirectional.bottomCenter,
             end: AlignmentDirectional.topCenter,
-            colors: <Color>[Colors.black45, Colors.transparent],
+            colors: <Color>[theme.dividerColor, Colors.transparent],
           ),
         ),
         child: Row(
           children: <Widget>[
-            Icon(Icons.videocam, size: 24.0),
+            const Icon(
+              Icons.videocam,
+              size: 24.0,
+              color: Colors.white,
+            ),
             Padding(
               padding: const EdgeInsets.only(left: 4.0),
               child: Text(
                 Constants.textDelegate.durationIndicatorBuilder(
-                    Duration(seconds: asset.duration)),
-                style: const TextStyle(fontSize: 16.0),
+                  Duration(seconds: asset.duration),
+                ),
+                style: const TextStyle(
+                  color: Colors.white,
+                  fontSize: 16.0,
+                ),
               ),
             ),
           ],
@@ -632,7 +665,9 @@ class AssetPicker extends StatelessWidget {
             },
             child: AnimatedContainer(
               duration: switchingPathDuration,
-              color: selected ? Colors.black45 : Colors.black.withOpacity(0.1),
+              color: selected
+                  ? theme.colorScheme.primary.withOpacity(0.45)
+                  : Colors.black.withOpacity(0.1),
             ),
           ), // 点击预览同目录下所有资源
         );
@@ -640,7 +675,7 @@ class AssetPicker extends StatelessWidget {
     );
   }
 
-  /// Indicator for asset selected status.
+  /// Indicator for assets selected status.
   /// 资源是否已选的指示器
   Widget _selectIndicator(AssetEntity asset) {
     return Selector<AssetPickerProvider, List<AssetEntity>>(
@@ -648,6 +683,7 @@ class AssetPicker extends StatelessWidget {
           provider.selectedAssets,
       builder: (BuildContext _, List<AssetEntity> selectedAssets, Widget __) {
         final bool selected = selectedAssets.contains(asset);
+        final double indicatorSize = Screens.width / gridCount / 3;
         return Positioned(
           top: 0.0,
           right: 0.0,
@@ -663,33 +699,43 @@ class AssetPicker extends StatelessWidget {
                 provider.selectAsset(asset);
               }
             },
-            child: AnimatedContainer(
-              duration: switchingPathDuration,
-              margin: EdgeInsets.all(isAppleOS ? 10.0 : 6.0),
-              width: isAppleOS ? 28.0 : 20.0,
-              height: isAppleOS ? 28.0 : 20.0,
-              decoration: BoxDecoration(
-                border: !selected
-                    ? Border.all(color: Colors.white, width: 2.0)
-                    : null,
-                color: selected ? themeColor : null,
-                shape: BoxShape.circle,
-              ),
-              child: AnimatedSwitcher(
+            child: Container(
+              margin: EdgeInsets.all(
+                  Screens.width / gridCount / (isAppleOS ? 12.0 : 15.0)),
+              width: indicatorSize,
+              height: indicatorSize,
+              alignment: AlignmentDirectional.topEnd,
+              child: AnimatedContainer(
                 duration: switchingPathDuration,
-                reverseDuration: switchingPathDuration,
-                child: selected
-                    ? isSingleAssetMode
-                        ? Icon(Icons.check, size: 18.0)
-                        : Text(
-                            '${selectedAssets.indexOf(asset) + 1}',
-                            style: TextStyle(
-                              fontSize: isAppleOS ? 16.0 : 14.0,
-                              fontWeight:
-                                  isAppleOS ? FontWeight.w600 : FontWeight.bold,
-                            ),
-                          )
-                    : const SizedBox.shrink(),
+                width: indicatorSize / (isAppleOS ? 1.25 : 1.5),
+                height: indicatorSize / (isAppleOS ? 1.25 : 1.5),
+                decoration: BoxDecoration(
+                  border: !selected
+                      ? Border.all(color: Colors.white, width: 2.0)
+                      : null,
+                  color: selected ? themeColor : null,
+                  shape: BoxShape.circle,
+                ),
+                child: AnimatedSwitcher(
+                  duration: switchingPathDuration,
+                  reverseDuration: switchingPathDuration,
+                  child: selected
+                      ? isSingleAssetMode
+                          ? Icon(Icons.check, size: 18.0)
+                          : Text(
+                              '${selectedAssets.indexOf(asset) + 1}',
+                              style: TextStyle(
+                                color: selected
+                                    ? theme.textTheme.bodyText1.color
+                                    : null,
+                                fontSize: isAppleOS ? 16.0 : 14.0,
+                                fontWeight: isAppleOS
+                                    ? FontWeight.w600
+                                    : FontWeight.bold,
+                              ),
+                            )
+                      : const SizedBox.shrink(),
+                ),
               ),
             ),
           ),
@@ -704,7 +750,7 @@ class AssetPicker extends StatelessWidget {
         child: Text(
           Constants.textDelegate.loadFailed,
           textAlign: TextAlign.center,
-          style: TextStyle(color: Colors.white, fontSize: 18.0),
+          style: const TextStyle(fontSize: 18.0),
         ),
       );
 
@@ -782,7 +828,7 @@ class AssetPicker extends StatelessWidget {
               gradient: LinearGradient(
                 begin: AlignmentDirectional.topCenter,
                 end: AlignmentDirectional.bottomCenter,
-                colors: <Color>[Colors.black45, Colors.transparent],
+                colors: <Color>[theme.dividerColor, Colors.transparent],
               ),
             ),
             child: Padding(
@@ -796,7 +842,7 @@ class AssetPicker extends StatelessWidget {
             ),
           ),
         ),
-        Center(child: Icon(Icons.audiotrack)),
+        const Center(child: Icon(Icons.audiotrack)),
         _selectedBackdrop(context, index, asset),
         audioIndicator(asset),
       ],
@@ -910,7 +956,9 @@ class AssetPicker extends StatelessWidget {
                           '(${provider.selectedAssets.length})'
                       : Constants.textDelegate.preview,
                   style: TextStyle(
-                    color: isSelectedNotEmpty ? null : Colors.grey[600],
+                    color: isSelectedNotEmpty
+                        ? null
+                        : theme.textTheme.caption.color,
                     fontSize: 18.0,
                   ),
                 );
@@ -977,7 +1025,7 @@ class AssetPicker extends StatelessWidget {
               )
             : IconButton(
                 onPressed: Navigator.of(context).maybePop,
-                icon: Icon(Icons.close),
+                icon: const Icon(Icons.close),
               ),
       );